+++ packages/visual-dev/src/components/Checkbox/Styles.ts
@@ -18,7 +18,7 @@ export const CheckboxLabelStyle = css`
   display: inline-flex;
   cursor: pointer;
   color: var(--sq-text);
-  font-family: "Helvetica Neue", Helvetica, sans-serif;
+  font-family: var(--sq-font-family-sans);
   font-size: var(--sq-font-size-regular);
   font-weight: var(--sq-font-weight-regular);
   line-height: var(--sq-line-height-regular);