+++ packages/visual-dev/src/components/TextLink/Styles.ts
@@ -1,0 +1,10 @@
+import { css } from "styled-components";
+
+export const base = css`
+  text-decoration: none;
+  color: #14b1f7;
+  cursor: pointer;
+  &:hover {
+    color: #14b1f7;
+  }
+`;